refactor: convert Results class to the FC
- Remove index file for re-export only for this file
- Fix path to the file in test file
