--- src/components/Results/index.ts
@@ -1,1 +1,0 @@
-export { Results } from './Results';
